happykh/requirements.txt
astroid==2.0.4
atomicwrites==1.2.0
attrs==18.1.0
coverage==4.5.1
Django==2.1.2
django-cors-headers==2.4.0
django-sendgrid-v5==0.6.893
djangorestframework==3.8.2
future==0.16.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
more-itertools==4.3.0
Pillow==5.2.0
pluggy==0.7.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
py==1.6.0
pylint==2.1.1
pytest==3.7.3
pytest-cov==2.5.1
pytest-django==3.4.2
python-http-client==3.1.0
pytz==2018.5
sendgrid==5.6.0
six==1.11.0
typed-ast==1.1.0
wrapt==1.10.11
